moving to cpp file to avoid longer compile times
